refactor: EnterCouponContent 내부 로직 hook으로 분리
